Add travis and appveyor for prebuilding binaries
