+++ README.md
@@ -11,9 +11,16 @@ Future plans include surfacing an API for running [Compute Shaders](https://www.
 
 ## Supported platforms
 
-* Mac OS
-* Windows
-* Linux
+This is a native Addon to NodeJS which delivers prebuilt binaries. Only some environments are supported:
+
+| Node Version       | Windows 64-Bit | Windows 32-Bit | Linux 64-Bit | Linux 32-Bit | OSX |
+| ------------------ | -------------- | -------------- | ------------ | ------------ | --- |
+| Earlier            | ✗              | ✗              | ✗            | ✗            | ✗   |
+| Node 10 _(Abi 64)_ | ✓              | ✗              | ✓            | ✗            | ✓   |
+| Node 11 _(Abi 67)_ | ✓              | ✗              | ✓            | ✗            | ✓   |
+| Node 12 _(Abi 72)_ | ✓              | ✗              | ✓            | ✗            | ✓   |
+| Node 13 _(Abi 79)_ | ✓              | ✗              | ✓            | ✗            | ✓   |
+| Node 14 _(Abi 83)_ | ✓              | ✗              | ✓            | ✗            | ✓   |
 
 ## Creating a WebGL context
