+++ package.json
@@ -15,7 +15,9 @@
   "scripts": {
     "build-npm": "./scripts/build-npm.sh",
     "format": "clang-format -i -style=Google binding/*.cc binding/*.h",
-    "install": "node scripts/install.js"
+    "install": "node scripts/install.js",
+    "build": "node scripts/build.js && tsc -p .",
+    "postbuild": "node scripts/rename.js"
   },
   "devDependencies": {
     "@types/bindings": "^1.3.0",
@@ -32,6 +34,7 @@
     "bindings": "^1.3.0",
     "https-proxy-agent": "^2.2.3",
     "progress": "^2.0.3",
+    "request": "^2.88.2",
     "rimraf": "^2.6.2",
     "tar": "^6.0.1"
   }
