+++ scripts/file-name.d.ts
@@ -1,0 +1,3 @@
+export const suffix: string;
+export const baseName: string;
+export const qualifiedName: string;
